Bugfix in analytics snippet
Fixes a bug where the analytics ids could only be found when you viewed a top-level-page. On pages in a folder/navigation or subpages, the id could not be found.

Now we're getting all inherited properties (from the site down to the page we're viewing) and checking if the ua & gtm ids are found there.